refactor: move pluginworkload metrics to the internal metrics package
On-behalf-of: @SAP krzysztof.zagorski@sap.com
